Cutscene over event
- Modders can now do something when a player stops watching a cutscene for some reason
- Cutscene length is now calculated more accurately
- Cutscene length no longer has to be an integer

+++ common/src/main/java/net/thewinnt/cutscenes/transition/NoopTransition.java
@@ -10,7 +10,7 @@ import net.thewinnt.cutscenes.CutsceneType;
 
 public class NoopTransition implements Transition {
     @Override
-    public int getLength() {
+    public double getLength() {
         return 0;
     }
 
@@ -35,12 +35,12 @@ public class NoopTransition implements Transition {
     }
     
     @Override
-    public int getOffCutsceneTime() {
+    public double getOffCutsceneTime() {
         return 0;
     }
 
     @Override
-    public int getOnCutsceneTime() {
+    public double getOnCutsceneTime() {
         return 0;
     }
